fix(tests): governance findings — split oversized test files, fix vacuous assertions (SMI-2755)
Fixes all 16 governance findings from Wave 2 code review:
- Critical: split recommend.test.ts (621→362) → recommend-online-path.test.ts (282)
- Critical: split search.test.ts (555→418) → search-online-path.test.ts (173)
- Critical: rename misleading 'rethrows' test in schema-migrations (asserted inverse)
- High: add expect.hasAssertions() to 4 vacuous conditional tests
- High: fix toBeGreaterThanOrEqual(0) → toBeGreaterThan(0) in 2 search tests
- High: replace 8 console.log guards with it.skipIf() in schema-migrations
- Medium: add SkillsmithError type to 3 bare toThrow() calls in compare
- Medium: add comments to 5 'as unknown as' partial mocks in recommend
- Medium: add pre-cast assertion in quota.test.ts
- Medium: move offline-path tracking test out of online describe block
- Low: remove redundant MCPTrustTier cast, strengthen toBeTruthy, raise timing threshold
