+++ packages/core/tests/db/schema-migrations.test.ts
@@ -19,6 +19,10 @@ import {
 import type { Database } from '../../src/db/database-interface.js'
 import { isBetterSqlite3Available } from '../../src/db/drivers/betterSqlite3Driver.js'
 
+// Evaluated once at module load — used by it.skipIf() to skip tests
+// when the SQLite driver is unavailable (e.g. WASM fallback mode).
+const skipIfNoSqlite = !isBetterSqlite3Available()
+
 const testDatabases: Database[] = []
 
 afterEach(() => {
@@ -42,60 +46,49 @@ async function freshDb(): Promise<Database> {
 // ============================================================================
 
 describe('runMigrationsSafe', () => {
-  it('handles "duplicate column" SQLite error gracefully — does not throw', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
+  it.skipIf(skipIfNoSqlite)(
+    'handles "duplicate column" SQLite error gracefully — does not throw',
+    async () => {
+      const db = await freshDb()
+
+      // createDatabaseAsync already ran initializeSchema which includes all columns.
+      // Running migrations again should encounter "duplicate column" on version 2/4
+      // and handle it silently — no throw expected.
+      expect(() => runMigrationsSafe(db)).not.toThrow()
     }
+  )
 
-    const db = await freshDb()
+  it.skipIf(skipIfNoSqlite)(
+    'does not throw when migrations are re-applied (idempotent)',
+    async () => {
+      const db = await freshDb()
 
-    // createDatabaseAsync already ran initializeSchema which includes all columns.
-    // Running migrations again should encounter "duplicate column" on version 2/4
-    // and handle it silently — no throw expected.
-    expect(() => runMigrationsSafe(db)).not.toThrow()
-  })
+      // Manually set schema_version to 0 so migration 1 runs again
+      db.exec('DELETE FROM schema_version')
+      db.prepare('INSERT INTO schema_version (version) VALUES (?)').run(0)
 
-  it('re-throws non-duplicate-column errors from migrations', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
+      // Running migration 1 again will try to CREATE TABLE skills which already exists.
+      // Migration 1 uses CREATE TABLE IF NOT EXISTS so it's idempotent — this is fine.
+      // What we want to test is that genuine errors DO propagate.
+      // We patch: inject a failing migration by checking internal migration behavior.
+      // Since we cannot inject a broken SQL path easily without monkey-patching MIGRATIONS,
+      // we verify the inverse: migrations succeed gracefully for the known schema.
+      expect(() => runMigrationsSafe(db)).not.toThrow()
     }
+  )
 
-    const db = await freshDb()
+  it.skipIf(skipIfNoSqlite)(
+    'handles FTS5 migration failure gracefully — does not throw',
+    async () => {
+      const db = await freshDb()
 
-    // Manually set schema_version to 0 so migration 1 runs again
-    db.exec('DELETE FROM schema_version')
-    db.prepare('INSERT INTO schema_version (version) VALUES (?)').run(0)
-
-    // Running migration 1 again will try to CREATE TABLE skills which already exists.
-    // Migration 1 uses CREATE TABLE IF NOT EXISTS so it's idempotent — this is fine.
-    // What we want to test is that genuine errors DO propagate.
-    // We patch: inject a failing migration by checking internal migration behavior.
-    // Since we cannot inject a broken SQL path easily without monkey-patching MIGRATIONS,
-    // we verify the inverse: migrations succeed gracefully for the known schema.
-    expect(() => runMigrationsSafe(db)).not.toThrow()
-  })
-
-  it('handles FTS5 migration failure gracefully — does not throw', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
+      // FTS5 table already exists from schema init. Running runMigrationsSafe
+      // calls db.exec(FTS5_MIGRATION_SQL) which attempts INSERT OR IGNORE — safe.
+      expect(() => runMigrationsSafe(db)).not.toThrow()
     }
+  )
 
-    const db = await freshDb()
-
-    // FTS5 table already exists from schema init. Running runMigrationsSafe
-    // calls db.exec(FTS5_MIGRATION_SQL) which attempts INSERT OR IGNORE — safe.
-    expect(() => runMigrationsSafe(db)).not.toThrow()
-  })
-
-  it('returns migration count run (0 when already up-to-date)', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
-    }
-
+  it.skipIf(skipIfNoSqlite)('returns migration count run (0 when already up-to-date)', async () => {
     const db = await freshDb()
 
     // Already at current schema version after createDatabaseAsync
@@ -104,12 +97,7 @@ describe('runMigrationsSafe', () => {
     expect(count).toBe(0) // All migrations already applied
   })
 
-  it('runs pending migrations and records version increments', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
-    }
-
+  it.skipIf(skipIfNoSqlite)('runs pending migrations and records version increments', async () => {
     const db = await freshDb()
 
     // Pin version to 1 (initial migration) to force re-running pending ones
@@ -134,12 +122,7 @@ describe('runMigrationsSafe', () => {
 // ============================================================================
 
 describe('closeDatabase', () => {
-  it('closes an open database without error', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
-    }
-
+  it.skipIf(skipIfNoSqlite)('closes an open database without error', async () => {
     const db = await createDatabaseAsync(':memory:')
     // Do NOT push to testDatabases — we close it manually
 
@@ -154,42 +137,38 @@ describe('closeDatabase', () => {
 // ============================================================================
 
 describe('createDatabaseAsync - schema initialization', () => {
-  it('creates schema_version table and sets to current version', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
+  it.skipIf(skipIfNoSqlite)(
+    'creates schema_version table and sets to current version',
+    async () => {
+      const db = await freshDb()
+
+      const version = getSchemaVersion(db)
+      expect(version).toBe(SCHEMA_VERSION)
     }
+  )
 
-    const db = await freshDb()
+  it.skipIf(skipIfNoSqlite)(
+    'initializes all required tables including schema_version',
+    async () => {
+      const db = await freshDb()
 
-    const version = getSchemaVersion(db)
-    expect(version).toBe(SCHEMA_VERSION)
-  })
-
-  it('initializes all required tables including schema_version', async () => {
-    if (!isBetterSqlite3Available()) {
-      console.log('Skipping test: no SQLite driver available')
-      return
-    }
-
-    const db = await freshDb()
-
-    const tables = db
-      .prepare(
-        `SELECT name FROM sqlite_master
+      const tables = db
+        .prepare(
+          `SELECT name FROM sqlite_master
          WHERE type='table' AND name NOT LIKE 'sqlite_%'
          ORDER BY name`
-      )
-      .all() as { name: string }[]
+        )
+        .all() as { name: string }[]
 
-    const tableNames = tables.map((t) => t.name)
-    expect(tableNames).toContain('schema_version')
-    expect(tableNames).toContain('skills')
-    expect(tableNames).toContain('sources')
-    expect(tableNames).toContain('categories')
-    expect(tableNames).toContain('cache')
-    expect(tableNames).toContain('audit_logs')
-  })
+      const tableNames = tables.map((t) => t.name)
+      expect(tableNames).toContain('schema_version')
+      expect(tableNames).toContain('skills')
+      expect(tableNames).toContain('sources')
+      expect(tableNames).toContain('categories')
+      expect(tableNames).toContain('cache')
+      expect(tableNames).toContain('audit_logs')
+    }
+  )
 })
 
 // ============================================================================
